+++ src/details/ArborX_Point.hpp
@@ -25,11 +25,11 @@ private:
   {
     constexpr Data() noexcept : coords{0.,0.,0.}{}
 
-    constexpr Data(const std::initializer_list<double> vals) noexcept :
+    constexpr Data(const std::initializer_list<float> vals) noexcept :
     coords {vals.begin()[0], vals.begin()[1], vals.begin()[2]}
     {}
 
-    double coords[3];
+    float coords[3];
   } _data = {};
 
 public:
@@ -55,22 +55,22 @@ public:
   }
 
   KOKKOS_INLINE_FUNCTION
-  constexpr double &operator[](unsigned int i) { return _data.coords[i]; }
+  constexpr float &operator[](unsigned int i) { return _data.coords[i]; }
 
   KOKKOS_INLINE_FUNCTION
-  constexpr const double &operator[](unsigned int i) const
+  constexpr const float &operator[](unsigned int i) const
   {
     return _data.coords[i];
   }
 
   KOKKOS_INLINE_FUNCTION
-  double volatile &operator[](unsigned int i) volatile
+  float volatile &operator[](unsigned int i) volatile
   {
     return _data.coords[i];
   }
 
   KOKKOS_INLINE_FUNCTION
-  double const volatile &operator[](unsigned int i) const volatile
+  float const volatile &operator[](unsigned int i) const volatile
   {
     return _data.coords[i];
   }
